requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.20
certifi==2015.11.20.1
Django==1.8
django-celery==3.1.17
django-haystack==2.4.1
elasticsearch==1.9.0
kombu==3.0.33
pyelasticsearch==1.4
pytz==2015.7
simplejson==3.8.1
six==1.10.0
SQLAlchemy==1.0.11
urllib3==1.14
wheel==0.24.0
